requirements.txt
django
djangorestframework
pyjwt
beautifulsoup4
requests
black
